Ahaaaaaa I understand more about how this all works!  And what invoked these .jsp files!
sdlfkdjsfldskfjdslfj good lord this is insane (to my subjective
preference)!  How many times does control flow go back and forth between
the webapp and Tomcat in a single page!? X'D  Look at how hard it is to
tell What Invokes This for any JSP page! X'D

Some are invoked by the main non-template pages (which were JSP files of
only Java code X"D), some are invoked by tags like <wiki:Content/> and
<wiki:Editor/> which pull the knowledge of which one from hidden
variables not passed to them X'D, and some are TOPLEVEL PAGES EXPOSED TO
THE INTERNET, but it takes hours to figure out which one is which!! X'D

We are definitely going to fix all of this so it has a monotonic call
structure XD
(or whatever that's called X3 )

Tomcat → Webapp → ActionBeans → JSP → TagFiles [→ TagFiles etc.
recursion]
(or Tomcat → Webapp → ActionBeans and other things like static images,
or machine-readable data generated right there without the need for
templating languages, or Tomcat → Webapp → WikiAjaxlets ..which we'll
probably just eventually turn into ActionBeans-that-don't-use-JSP also,
for simplicity X3 )

And that's it and always it XD